+++ Source/SwiftCarousel.swift
@@ -85,7 +85,8 @@ public class SwiftCarousel: UIView {
         
         return index
     }
-    /// Returns carousel views.
+    /// Carousel items. You can setup your carousel using this method (static items), or
+    /// you can also see `itemsFactory`, which uses closure for the setup.
     public var items: [UIView] {
         get {
             return [UIView](choices[choices.count / 3..<(choices.count / 3 + originalChoicesNumber)])
@@ -98,13 +99,43 @@ public class SwiftCarousel: UIView {
                     if counter == 1 {
                         return choice
                     } else {
-                        return choice.copyView()
+                        do {
+                            return try choice.copyView()
+                        } catch {
+                            fatalError("There was a problem with copying view.")
+                        }
                     }
                 }
                 self.choices.appendContentsOf(newViews)
             }
             setupViews(choices)
         }
+    }
+    
+    /// Factory for carousel items. Here you specify how many items do you want in carousel
+    /// and you need to specify closure that will create that view. Remember that it should
+    /// always create new view, not give the same reference all the time.
+    /// You can always setup your carousel using `items` instead.
+    public func itemsFactory(itemsCount count: Int, factory: (index: Int) -> UIView) {
+        guard count > 0 else { return }
+        
+        originalChoicesNumber = count
+        (0..<3).forEach { counter in
+            let newViews: [UIView] = 0.stride(to: count, by: 1).map { i in
+                var view = factory(index: i)
+                if self.choices.contains(view) {
+                    do {
+                        view = try view.copyView()
+                    } catch {
+                        fatalError("There was a problem with copying view.")
+                    }
+                }
+                
+                return view
+            }
+            self.choices.appendContentsOf(newViews)
+        }
+        setupViews(choices)
     }
     
     // MARK: - Inits
@@ -125,10 +156,10 @@ public class SwiftCarousel: UIView {
      - parameter choices: Items to put in carousel.
      
      */
-    public convenience init(frame: CGRect, choices: [UIView]) {
+    public convenience init(frame: CGRect, items: [UIView]) {
         self.init(frame: frame)
         setup()
-        items = choices
+        self.items = items
     }
     
     deinit {